Color code action text depending on task business status
