+++ opensrp-reveal/src/main/java/org/smartregister/reveal/viewholder/TaskRegisterViewHolder.java
@@ -9,6 +9,7 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import org.smartregister.reveal.R;
+import org.smartregister.reveal.model.BaseCardDetails;
 import org.smartregister.reveal.model.TaskDetails;
 
 /**
@@ -54,8 +55,15 @@ public class TaskRegisterViewHolder extends RecyclerView.ViewHolder {
         distanceView.setVisibility(View.GONE);
     }
 
-    public void setTaskAction(String actionLabel, TaskDetails task, View.OnClickListener onClickListener) {
+    public void setTaskAction(String actionLabel, TaskDetails task, BaseCardDetails cardDetails, View.OnClickListener onClickListener) {
         actionView.setText(actionLabel);
+        if (cardDetails != null && cardDetails.getStatusColor() != null) {
+            actionView.setBackground(null);
+            actionView.setTextColor(context.getResources().getColor(cardDetails.getStatusColor()));
+        } else {
+            actionView.setBackground(context.getResources().getDrawable(R.drawable.task_action_bg));
+            actionView.setTextColor(context.getResources().getColor(R.color.text_black));
+        }
         actionView.setOnClickListener(onClickListener);
         actionView.setTag(R.id.task_details, task);
     }
