+++ opensrp-reveal/src/main/res/layout/task_register_row.xml
@@ -38,6 +38,7 @@
         android:layout_height="wrap_content"
         android:paddingStart="@dimen/task_register_padding"
         android:paddingEnd="@dimen/task_register_padding"
+        android:paddingBottom="@dimen/task_register_padding"
         android:textColor="#7A7A7A"
         android:textSize="@dimen/task_register_text_size"
         app:layout_constraintStart_toEndOf="@id/task_icon"
